Move Size and Color to Colore.Core.
