+++ Colore/Colore.csproj
@@ -76,8 +76,9 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="ColoreException.cs" />
+    <Compile Include="Core\Color.cs" />
     <Compile Include="Core\Colore.cs" />
-    <Compile Include="Razer\Color.cs" />
+    <Compile Include="Core\Size.cs" />
     <Compile Include="Razer\Constants.cs" />
     <Compile Include="Razer\Keyboard\Constants.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
@@ -100,7 +101,6 @@
     <Compile Include="Razer\NativeMethods.cs" />
     <Compile Include="Razer\Product.cs" />
     <Compile Include="Razer\Result.cs" />
-    <Compile Include="Razer\Size.cs" />
   </ItemGroup>
   <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
