+++ Colore/Core/Size.cs
@@ -28,11 +28,10 @@
 // </copyright>
 // ---------------------------------------------------------------------------------------
 
-namespace Colore.Razer
+namespace Colore.Core
 {
     using System;
     using System.Runtime.InteropServices;
-
 #if WIN64
 
     using size_t = System.UInt64;
